+++ usecase/post.go
@@ -2,6 +2,7 @@ package usecase
 
 import (
 	"context"
+	"fmt"
 	"mikke-server/database"
 	"mikke-server/domain"
 )
@@ -22,4 +23,17 @@ func (p *PostUsecase) PostQuestion(ctx context.Context, user_id int, title strin
 		return nil, err
 	}
 	return post, nil
+}
+
+type ListPostsUsecase struct {
+	Repo PostLister
+	DB   database.Queryer
+}
+
+func (p *ListPostsUsecase) ListPosts(ctx context.Context) (domain.Posts, error) {
+	posts, err := p.Repo.ListPosts(ctx, p.DB)
+	if err != nil {
+		return nil, fmt.Errorf("failed to list: %w", err)
+	}
+	return posts, nil
 }
